change example in readme - that syntax doesnt work
readme syntax for qless.new() redis config doesnt work - it results in the defaults being used (this might not be obvious if your redis is localhost..).   extrapolating from ledge's use for qless_worker.new(), I was able to get it to connect - removed the redis{} section around the host/port/password/socket section.